fitness_webapp/requirements.txt
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.27
werkzeug==2.0.2
pandas
matplotlib
seaborn
scikit-learn
wordcloud
nltk
joblib
Pillow
Flask-Session
python-dotenv 